% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate.asymmetric.G}
\alias{generate.asymmetric.G}
\title{Generate an unweighted asymmetric interaction matrix from a graph}
\usage{
generate.asymmetric.G(graph)
}
\arguments{
\item{graph}{igraph Graph object}
}
\value{
a generated asymmetric interaction matrix
}
\description{
An unweighted asymmetric interaction matrix, denoted as (G), to present
underlying type of taxa interactions as either commensalistic (0),
competitive (-1), and mutalistic (+1).
This function randomly rewrites edges denoted in an adjacency
matrix by randomly switch some (+1) to (-1).
}
\examples{
g = igraph::sample_gnm(n=5, m=5)
unweighted_interaction_matrix = SPLANG:::generate.asymmetric.G(graph=g)
unweighted_interaction_matrix
}
