% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{get.coupled.mu.Omega}
\alias{get.coupled.mu.Omega}
\title{Calculate coupled mean abundance and a precision matrix}
\usage{
get.coupled.mu.Omega(G, u, v)
}
\arguments{
\item{G}{underlying unweighted asymmetric interaction matrix}

\item{u}{offset parameter}

\item{v}{scale parameter}
}
\value{
a named list of coupled expected abundance (mu) and precision matrix (Omega)
}
\description{
Calculate coupled mean abundance (mu) and a precision matrix (Omega) from
the underlying unweighted asymmetric interaction matrix (G).
}
\examples{
graph = igraph::sample_gnm(n=5, m=5) 
G = SPLANG:::generate.asymmetric.G(graph)
coupled.mu.Omega = SPLANG:::get.coupled.mu.Omega(G, u=0.1, v=0.3)
coupled.mu.Omega
}
