% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{sample.sequencing}
\alias{sample.sequencing}
\title{Generate abundance samples and underlying taxa interaction network}
\usage{
sample.sequencing(
  n.species,
  n.sequencing,
  sequencing.depth,
  u = 0.1,
  v = 0.3,
  network.type = c("nm", "pa"),
  m.edges = round(n.species * (n.species - 1) * 0.8),
  power = 1
)
}
\arguments{
\item{n.species}{number of taxa, an integer.}

\item{n.sequencing}{number of sequencing performed, an integer.}

\item{sequencing.depth}{depth of each sequencing, an integer.}

\item{u}{offset parameter, a float number.}

\item{v}{scale parameter, a non-negative float number.}

\item{network.type}{type of underlying network, either
preferential attaching ("pa"), or Erdos-Renyi ("nm")}

\item{m.edges}{number of edges if network type is Erdos-Renyi N-M,
the default is \code{round(n.species * (n.species - 1) *.8)},
i.e. randomly choose 80\% of all possible edges
only used if \code{network.type="nm"}. The \code{floor(m.edge)} will be used if
\code{m.edge} is not an integer.}

\item{power}{the power of preferential attachment, the default is one,
i.e. linear preferential attachment.
This parameter is meaningful only when the \code{network.type="pa"}, and it will
be ignored otherwise.}
}
\value{
a named-list of:\if{html}{\out{<br>}}
\itemize{
\item underlying unweighted interaction network (\code{G}) with shape \code{n.species} \eqn{\times}{x} \code{n.species};\if{html}{\out{<br>}}
\item partial correlation matrix (\code{omega}) with shape \code{n.species} \eqn{\times}{x} \code{n.species};\if{html}{\out{<br>}}
\item ground truth abundance (\code{groundtruth.abundance}) with shape 1 \eqn{\times}{x} \code{n.species};\if{html}{\out{<br>}}
\item synthetic sequencing data (\code{synthetic.sequencing.data}) as a matrix with
\code{n.species} rows and \code{n.sequencing} columns, with each row representing
\code{n.sequencing} results of the i-th species. The result is a counting ranging
from 0 to \code{sequencing.depth}.
}
}
\description{
This function will firstly generate an underlying taxa interaction network,
then generate synthetic abundance samples with specified number of sequencings
performed and the depth of each sequencing.
}
\examples{
# generate abundance sequencing for preferential-attaching typed underlying network
sample.sequencing(5, 10, 100, network.type = 'pa')

# generate abundance sequencing for Erdo-Renyi typed underlying network with 
# 80\% of all possible edges
sample.sequencing(5, 10, 100, network.type = 'nm')

# generate abundance sequencing for Erdo-Renyi typed underlying network with 
# 50\% of all possible edges
sample.sequencing(
  n.species=5, 
  n.sequencing=10, 
  sequencing.depth=100, 
  network.type = 'nm',
  m.edges=(5 * (5 - 1) * .5) # use the floor value if not an integer
  )

}
